auth features updates. form components for register, login & reset password feature created in components/auth. privacy policy and t&c checkoboxes/links added to the register page. Global component input-checkbox created

+++ app/Http/Controllers/BookmarkController.php
@@ -22,12 +22,16 @@ class BookmarkController extends Controller
     }
 
     /**
-     * Add - Remove bookmark.
+     * Add - Remove bookmark
      */
     public function bookmark(CarListing $listing): RedirectResponse
     {
         $user = Auth::user();
 
+        // redirect user user if user is owner
+        if($user->id == $listing->user_id) return back();
+
+        // if user is not owner
         if ($user->userBookmarks()->where('car_listing_id', $listing->id)->exists()) {
             $user->userBookmarks()->detach($listing->id);
